requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amazon-textract-response-parser==1.0.3
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.51.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.2.0
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
boto3==1.35.42
botocore==1.35.42
build==1.2.2.post1
cachetools==5.3.3
celery==5.5.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cryptography==43.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.115.9
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-genai==1.7.0
google-resumable-media==2.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.6
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
kubernetes==32.0.1
langchain==0.3.25
langchain-anthropic==0.3.13
langchain-chroma==0.2.3
langchain-community==0.3.23
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langserve==0.3.1
langsmith==0.3.42
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.1.0
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.77.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.0
passlib==1.7.4
pdf2image==1.17.0
pillow==10.2.0
pluggy==1.4.0
posthog==4.0.1
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.23.0
protobuf==5.29.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
PyExecJS==1.5.1
Pygments==2.19.1
pyinstrument==4.6.2
PyJWT==2.9.0
pymongo==4.10.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.3.0
python-multipart==0.0.9
python-pptx==0.6.23
python-socketio==5.11.4
PyYAML==6.0.2
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
s3transfer==0.10.3
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.0.1
tqdm==4.66.2
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
wheel==0.44.0
wrapt==1.17.2
wsproto==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0